added grocery domain and test for merge

+++ generators/post_experiments.py
@@ -20,11 +20,23 @@ def summarize_csv(exp_dir):
         if len(rows) > 2:
 
             ## make header shorter
-            headers = [s.replace('timestamp', 't').replace('_count', '').replace('_time', '').replace('_length', '').replace('translation', 'parse').replace('_expanded', '') for s in rows[0]]
+            headers = [s.replace('timestamp', 't').replace('problem_name', 'run_name').replace('_count', '').replace('_time', '').replace('_length', '').replace('plan_cost', 'cost').replace('translation', 'parse').replace('_expanded', '') for s in rows[0]]
             rows = rows[1:]
 
             ## for problem_name, print only the difference
-            keys = [row[1].replace('/', '-').split('-') for row in rows]
+            keys = []
+            for row in rows:
+                key = row[1]
+                key = key[key.index('/')+1:]
+                key = key[:key.index('/')]
+                keys.append(key.split('-'))
+
+            ## pad optional fields like action_sub and goal_ignore with ''
+            max_len = max([len(key) for key in keys])
+            for key in keys:
+                while len(key) < max_len:
+                    key.append('')
+
             new_keys = ['' for i in range(len(keys))]
             for i in range(len(keys[0])):
                 col = [key[i] for key in keys]
@@ -35,10 +47,8 @@ def summarize_csv(exp_dir):
             for line in new_keys:
                 line = line[1:]
                 key = line[line.index('-')+1:]
-                if len(key) - len(key.replace('-','')) == 1:
-                    A, B = key.split('-')
-                    if A in B:
-                        key = B
+                key = key.replace('.pddl','').replace('-',', ')
+                key = key.replace('dmn=','').replace('prb=','').replace('obj=','')
                 row_names.append(key)        
             
             for i in range(len(rows)):
@@ -54,15 +64,17 @@ def save_copy(exp_dir, save_dir):
             replacing the previous successful save if there exists,
         also copy the test shell script """
 
+    if args.exp_sh == None: return
     exp_name = exp_dir[exp_dir.rfind('/')+1:]
-    dt = datetime.now().strftime("%m%d-%H%M%S")
-    save_dir = join(save_dir, f'{exp_name}-{dt}')
 
     ## delete all previous experiments with this name
     if args.replace:
         previous_dirs = [join(save_dir, f) for f in listdir(save_dir) if f"{exp_name}-" in f]
         for each in previous_dirs:
             shutil.rmtree(each)
+            
+    dt = datetime.now().strftime("%m%d-%H%M%S")
+    save_dir = join(save_dir, f'{exp_name}-{dt}')
 
     runs = [join(exp_dir, f) for f in listdir(exp_dir) if isdir(join(exp_dir, f))]
     failed_runs = len(runs)
@@ -71,7 +83,7 @@ def save_copy(exp_dir, save_dir):
         if isfile(plan_file) and getsize(plan_file) > 0:
             failed_runs -= 1
     
-    if failed_runs == 0:
+    if failed_runs == 0 or True:
         shutil.copytree(exp_dir, save_dir) 
         if args.exp_sh != None:
             shutil.copy(args.exp_sh, join(save_dir, args.exp_sh))
